+++ src/app/Http/Requests/AccountInfoRequest.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Backpack\Base\app\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
+
+class AccountInfoRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        // only allow updates if the user is logged in
+        return Auth::check();
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        $user = Auth::user();
+
+        return [
+            'email' => [
+                'required',
+                'email',
+                Rule::unique($user->getTable())->ignore($user->getKey()),
+            ],
+            'name' => 'required',
+        ];
+    }
+}