+++ src/resources/views/auth/account/sidemenu.blade.php
@@ -1,0 +1,17 @@
+<div class="box">
+	<ul class="nav nav-pills nav-stacked">
+
+	  <li role="presentation"
+		@if (Request::route()->getName() == 'backpack.account.info')
+	  	class="active"
+	  	@endif
+	  	><a href="{{ route('backpack.account.info') }}">{{ trans('backpack::base.update_account_info') }}</a></li>
+
+	  <li role="presentation"
+		@if (Request::route()->getName() == 'backpack.account.password')
+	  	class="active"
+	  	@endif
+	  	><a href="{{ route('backpack.account.password') }}">{{ trans('backpack::base.change_password') }}</a></li>
+
+	</ul>
+</div>